Get 1 customer and delete work

+++ flight-manager/images/mysql/data/flight_manager.sql
@@ -18,7 +18,7 @@ CREATE TABLE customer (
   customer_pseudo VARCHAR(45) NOT NULL,
   firstname VARCHAR(20) NOT NULL,
   lastname VARCHAR(20) NOT NULL,
-  age SMALLINT UNSIGNED NOT NULL,
+  age INT UNSIGNED NOT NULL,
   customer_pw VARCHAR(200) NOT NULL,
   PRIMARY KEY  (customer_id)
 )ENGINE=InnoDB DEFAULT CHARSET=utf8;
